backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
python-dotenv==1.0.1
firebase-admin==6.6.0
openai==1.57.2
anthropic==0.39.0
pydantic==2.12.0
python-multipart==0.0.18
spacy==3.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-cloud-secret-manager==2.21.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1
pytest-mock==3.14.0
structlog==25.5.0
google-cloud-error-reporting==1.12.0
